score_clusterisation_rag/requirements.txt
sentence-transformers==3.0.1
bertopic==0.16.3
datasets==2.20.0
accelerate==0.33.0
bitsandbytes==0.43.3
xformers==0.0.27.post2
adjustText==1.2.0
hdbscan==0.8.38.post1
umap-learn==0.5.6
plotly==5.22.0
scikit-learn==1.5.1
pandas==2.2.2
angchain==0.2.7
langchain_community==0.2.7
chromadb==0.5.4


